+++ src/main/java/roidrole/patternbanners/recipe/_Recipe.java
@@ -15,15 +15,27 @@ import roidrole.patternbanners.config.ConfigMapping;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
 
 import static roidrole.patternbanners.PatternBanners.pattern;
 
 public class _Recipe {
-    public static List<PatternApply> PATTERN_APPLY_RECIPES = new ArrayList<>();
+    public static List<PatternApply> PATTERN_APPLY_RECIPES;
     public static List<BannerPattern> PATTERNS_ONLY_SHAPE = new ArrayList<>();
 
     public static void init(){
-        if(ConfigGeneral.Patterns.craftable){
+        //Populate registries
+        if(ConfigGeneral.recipes.patternOnlyShape.enabled){
+            for(BannerPattern pattern : BannerPattern.values()){
+                if(pattern.hasPatternItem() || !pattern.hasPattern()){continue;}
+                PATTERNS_ONLY_SHAPE.add(pattern);
+            }
+        }
+        if(ConfigGeneral.recipes.patternApply.enabled){
+            PATTERN_APPLY_RECIPES = ConfigMapping.mappings.stream().map(PatternApply::new).collect(Collectors.toList());
+        }
+
+        if(ConfigGeneral.patterns.craftable){
             for (ConfigCategory mapping : ConfigMapping.mappings){
                 ResourceLocation name = new ResourceLocation(Tags.MOD_ID, "recipes/pattern_create/" + mapping.get("hash").getString());
                 if(mapping.containsKey("item") && !mapping.get("item").getString().isEmpty()){
@@ -34,26 +46,13 @@ public class _Recipe {
                             Ingredient.fromStacks(Utils.getItemStack(mapping))
                     );
                 } else if (mapping.containsKey("shap") && mapping.get("shap").getString().length() == 11) {
-                    PatternFromShape nextOne = new PatternFromShape(mapping);
-                    if(ConfigGeneral.Recipes.craftingTable) {
-                        GameData.register_impl(nextOne.setRegistryName(name));
-                    }
+                    GameData.register_impl(new PatternFromShape(mapping).setRegistryName(name));
                 }
             }
         }
 
-        if(!ConfigGeneral.Patterns.shapes_pattern){
-            for(BannerPattern pattern : BannerPattern.values()){
-                if(pattern.hasPatternItem() || !pattern.hasPattern()){continue;}
-                PATTERNS_ONLY_SHAPE.add(pattern);
-            }
-        }
-        for (ConfigCategory mapping : ConfigMapping.mappings) {
-            PatternApply nextOne = new PatternApply(mapping);
-            PATTERN_APPLY_RECIPES.add(nextOne);
-            if(ConfigGeneral.Recipes.craftingTable) {
-                GameData.register_impl(nextOne.setRegistryName(Tags.MOD_ID, "recipes/pattern_apply/" + mapping.get("hash").getString()));
-            }
+        if(ConfigGeneral.recipes.patternApply.craftingTable) {
+            PATTERN_APPLY_RECIPES.forEach(recipe -> GameData.register_impl(recipe.setRegistryName(Tags.MOD_ID, "recipes/pattern_apply/" + recipe.patternH)));
         }
     }
 }